refactor(theme): simplify the theme config
